+++ test/test_helper.exs
@@ -1,1 +1,2 @@
 ExUnit.start(exclude: [:interop, :hardware])
+Code.require_file("support/client.ex", __DIR__)